Add MecaDetails, to populate the session prior to ExpandMeca.
