+++ workflow/workflow_IngestMeca.py
@@ -35,6 +35,7 @@ class workflow_IngestMeca(Workflow):
             "start": {"requirements": None},
             "steps": [
                 define_workflow_step("PingWorker", data),
+                define_workflow_step("MecaDetails", data),
                 define_workflow_step("ExpandMeca", data),
                 define_workflow_step("ModifyMecaXml", data),
                 define_workflow_step("MecaPeerReviews", data),
